requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.6
easydict==1.13
exceptiongroup==1.3.0
fastapi==0.120.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
h11==0.16.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
insightface==0.7.3
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
onnx==1.19.1
onnxruntime-gpu==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==24.2
pillow==12.0.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
prettytable==3.16.0
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
stringzilla==4.2.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
yarl==1.22.0
